(FACT-1042) Allow time for EC2 to respond
Without this change, EC2 will not always respond to the metadata request
before the http session times out. Extend the session timeout to 5
seconds, and add a connection timeout of 200 ms to ensure Facter doesn't
spend too long waiting on the http request if not in EC2. Update the
error and add more detailed trace info in case EC2 is very slow to
respond.

Also extend the GCE session timeout to 5 seconds, and set a connection
timeout of 1 second to avoid long waits if GCE is unavailable.