(FACT-1016) Prevent unbounded growth of fact resolutions.
A bug in Puppet exposed an unbounded memory growth if the same custom
fact file is repeatedly loaded.  The reason is that every execution
results in a new resolution being added to the fact.  While this itself
is desired behavior, we should limit the number of resolutions to some
arbitrary size so that memory doesn't grow unbounded.